+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ## 2022/05/25
 - start of the project
+- started working on the lexer